Support .ttc files
When processing a .ttc file with multiple embedded fonts, it patches all
of them and generates a new .ttc. Presumably the added symbols won't
look right in the non-Regular variants, but it's better than omitting
them.

This has not been tested on non-TTC fonts. I assume that they'll only
contain one font, which means they'll behave as expected.